# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: S3MP API
  version: 1.1.0
  description: S3MP 是面向多租户的受控文件存储服务。管理接口使用租户会话 Cookie `s3mp_session`；账户登录与平台控制接口使用账户会话 Cookie `s3mp_account_session`。浏览器发起变更请求时同时提交
    `X-S3MP-CSRF`，服务端应用访问使用 API Key。错误响应统一返回稳定的 `code`、可读的 `message` 和用于排障的 `request_id`；任何会话值、API Key
    或存储凭据均不会出现在响应中。
paths:
...
  /api/v1/me:
    get:
      tags:
      - Context
      summary: 获取当前已选租户中的身份、成员关系与有效权限上下文。
      operationId: get_me
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '422': *id001
        '500': *id001
      description: 获取当前已选租户中的身份、成员关系与有效权限上下文。
  /api/v1/users:
    get:
      tags:
      - Context
      summary: 列出当前租户可见的用户账户。
      operationId: list_users
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/UserStatus'
          default: active
        description: 资源当前生命周期状态。
      - name: principal_type
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PrincipalType'
          default: user
        description: “principal_type”字段；具体取值和约束见当前资源说明。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出当前租户可见的用户账户。
      x-permission: members.read
  /api/v1/member-candidates:
    get:
      tags:
      - Context
      summary: 搜索尚未加入当前租户的全局已注册用户候选
      operationId: list_member_candidates
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberCandidatePage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 搜索尚未加入当前租户的全局已注册用户候选；支持邮箱、姓名和工号模糊查询。
      x-permission: members.manage
  /api/v1/users/{user_id}:
    get:
      tags:
      - Context
      summary: 获取当前租户中指定用户的公开身份信息。
      operationId: get_user
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: 全局用户账户的服务端唯一标识。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取当前租户中指定用户的公开身份信息。
      x-permission: members.read
  /api/v1/members:
    get:
      tags:
      - Context
      summary: 列出当前租户的成员关系。
      operationId: list_members
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/MembershipStatus'
          default: active
        description: 资源当前生命周期状态。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出当前租户的成员关系。
      x-permission: members.read
    post:
      tags:
      - Context
      summary: 将已有用户加入当前租户，或创建受邀成员关系。
      operationId: create_member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 将已有用户加入当前租户，或创建受邀成员关系。
      x-permission: members.manage
  /api/v1/members/{membership_id}:
    get:
      tags:
      - Context
      summary: 获取当前租户中指定成员关系的详情。
      operationId: get_member
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Membership Id
        description: 租户成员关系的服务端唯一标识。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取当前租户中指定成员关系的详情。
      x-permission: members.read
    patch:
      tags:
      - Context
      summary: 更新成员状态
      operationId: update_member
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Membership Id
        description: 租户成员关系的服务端唯一标识。
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
        description: 客户端已读取的资源 ETag；用于拒绝覆盖较新的服务端状态。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 更新成员状态；暂停、移除或到期会立即回收租户权限。
      x-permission: members.manage
  /api/v1/groups/{group_id}/members:
    get:
      tags:
      - Context
      summary: 列出指定用户组中的成员关系。
      operationId: list_group_members
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
        description: 用户组的服务端唯一标识。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出指定用户组中的成员关系。
      x-permission: groups.read
    post:
      tags:
      - Context
      summary: 将当前租户成员加入指定用户组。
      operationId: add_group_member
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
        description: 用户组的服务端唯一标识。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberBody'
      responses:
        '204':
          description: Successful Response
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 将当前租户成员加入指定用户组。
      x-permission: groups.manage
  /api/v1/groups/{group_id}/members/{membership_id}:
    delete:
      tags:
      - Context
      summary: 将成员从指定用户组移除。
      operationId: remove_group_member
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
        description: 用户组的服务端唯一标识。
      - name: membership_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Membership Id
        description: 租户成员关系的服务端唯一标识。
      responses:
        '204':
          description: Successful Response
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 将成员从指定用户组移除。
      x-permission: groups.manage
...
components:
  schemas:
...
    AddGroupMemberBody:
      properties:
        membership_id:
          type: string
          format: uuid
          title: Membership Id
          description: 租户成员关系的服务端唯一标识。
      additionalProperties: false
      type: object
      required:
      - membership_id
      title: AddGroupMemberBody
...
    MeResponse:
      properties:
        principal:
          $ref: '#/components/schemas/PrincipalSummary'
          description: 作为授权主体的账户、成员或服务身份摘要。
        current_tenant:
          $ref: '#/components/schemas/TenantSummary'
          description: 当前账户已选择的租户摘要；未选择时为 null。
        available_tenants:
          items:
            $ref: '#/components/schemas/TenantSummary'
          type: array
          title: Available Tenants
          description: 当前账户可选择的活跃租户摘要列表。
        coarse_permissions:
          items:
            type: string
          type: array
          uniqueItems: true
          title: Coarse Permissions
          description: 用于快速判定的粗粒度有效权限名称集合。
        authorization_version:
          type: integer
          minimum: 1.0
          title: Authorization Version
          description: 授权状态版本；变化后旧会话和缓存会被重新验证。
      additionalProperties: false
      type: object
      required:
      - principal
      - current_tenant
      - available_tenants
      - coarse_permissions
      - authorization_version
      title: MeResponse
    MemberCandidatePage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemberCandidateResponse'
          type: array
          title: Items
          description: 当前页返回的资源记录列表。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页分页游标；无更多结果时为 null。
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: MemberCandidatePage
    MemberCandidateResponse:
      properties:
        id:
          type: string
          title: Id
          description: 资源的服务端唯一标识。
        email:
          type: string
          format: email
          title: Email
          description: 用户登录邮箱地址。
        display_name:
          type: string
          title: Display Name
          description: 面向用户展示的名称。
      additionalProperties: false
      type: object
      required:
      - id
      - email
      - display_name
      title: MemberCandidateResponse
    MembershipCreate:
      properties:
        email:
          type: string
          maxLength: 320
          minLength: 3
          format: email
          title: Email
          description: 用户登录邮箱地址。
        display_name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Display Name
          description: 面向用户展示的名称。
        role_id:
          type: string
          format: uuid
          title: Role Id
          description: 角色的服务端唯一标识。
      additionalProperties: false
      type: object
      required:
      - email
      - role_id
      title: MembershipCreate
    MembershipPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MembershipResponse'
          type: array
          title: Items
          description: 当前页返回的资源记录列表。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页分页游标；无更多结果时为 null。
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: MembershipPage
    MembershipResponse:
      properties:
        id:
          type: string
          title: Id
          description: 资源的服务端唯一标识。
        user:
          $ref: '#/components/schemas/UserResponse'
          description: 全局用户账户的公开身份摘要。
        principal:
          $ref: '#/components/schemas/PrincipalSummary'
          description: 作为授权主体的账户、成员或服务身份摘要。
        status:
          type: string
          enum:
          - invited
          - active
          - suspended
          - removed
          title: Status
          description: 资源当前生命周期状态。
        authorization_version:
          type: integer
          minimum: 1.0
          title: Authorization Version
          description: 授权状态版本；变化后旧会话和缓存会被重新验证。
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 资源创建时间，采用 Asia/Shanghai（UTC+08:00）格式。
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: 资源最近更新时间，采用 Asia/Shanghai（UTC+08:00）格式。
        etag:
          type: string
          title: Etag
          description: 资源当前版本标识，用于并发控制。
      additionalProperties: false
      type: object
      required:
      - id
      - user
      - principal
      - status
      - authorization_version
      - created_at
      - updated_at
      - etag
      title: MembershipResponse
    MembershipStatus:
      type: string
      enum:
      - invited
      - active
      - suspended
      - removed
      title: MembershipStatus
    MembershipUpdate:
      properties:
        status:
          type: string
          enum:
          - invited
          - active
          - suspended
          - removed
          title: Status
          description: 资源当前生命周期状态。
        reason:
          type: string
          maxLength: 500
          minLength: 1
          title: Reason
          description: 执行该敏感操作的业务原因。
      additionalProperties: false
      type: object
      required:
      - status
      - reason
      title: MembershipUpdate
...
    PrincipalSummary:
      properties:
        id:
          type: string
          title: Id
          description: 资源的服务端唯一标识。
        type:
          type: string
          enum:
          - user
          - group
          - application
          title: Type
          description: 资源、错误或操作的类型标识。
        display_name:
          type: string
          maxLength: 200
          minLength: 1
          title: Display Name
          description: 面向用户展示的名称。
      additionalProperties: false
      type: object
      required:
      - id
      - type
      - display_name
      title: PrincipalSummary
    PrincipalType:
      type: string
      enum:
      - user
      - group
      - application
      title: PrincipalType
...
    TenantSummary:
      properties:
        id:
          type: string
          title: Id
          description: 资源的服务端唯一标识。
        name:
          type: string
          title: Name
          description: 资源名称，在所属租户范围内使用。
        membership_status:
          type: string
          enum:
          - invited
          - active
          - suspended
          - removed
          title: Membership Status
          description: 租户成员关系的当前状态。
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - membership_status
      title: TenantSummary
...
    UserPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserResponse'
          type: array
          title: Items
          description: 当前页返回的资源记录列表。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页分页游标；无更多结果时为 null。
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: UserPage
    UserResponse:
      properties:
        id:
          type: string
          title: Id
          description: 资源的服务端唯一标识。
        email:
          type: string
          format: email
          title: Email
          description: 用户登录邮箱地址。
        display_name:
          type: string
          title: Display Name
          description: 面向用户展示的名称。
        status:
          type: string
          enum:
          - active
          - disabled
          - deleted
          title: Status
          description: 资源当前生命周期状态。
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 资源创建时间，采用 Asia/Shanghai（UTC+08:00）格式。
      additionalProperties: false
      type: object
      required:
      - id
      - email
      - display_name
      - status
      - created_at
      title: UserResponse
    UserStatus:
      type: string
      enum:
      - active
      - disabled
      - deleted
      title: UserStatus
...