# src/s3mp/identity/api/router.py
"""Strict HTTP boundary for identity and membership management."""

from datetime import datetime
from typing import Annotated, Literal
from uuid import UUID

from fastapi import APIRouter, Depends, Header, Query, Response
from pydantic import BaseModel, ConfigDict, Field

from s3mp.common.api.cursor import CursorCodec
from s3mp.common.api.dependencies import (
    application_service,
    management_permission,
    principal_context,
)
from s3mp.common.api.etag import check_etag, require_if_match
from s3mp.identity.application.management_service import IdentityManagementService
from s3mp.identity.domain.context import PrincipalContext
from s3mp.identity.infrastructure.models import MembershipStatus, PrincipalType, UserStatus

router = APIRouter(prefix="/api/v1", tags=["Context"])
identity_service = application_service("identity_management")


class _Strict(BaseModel):
    model_config = ConfigDict(extra="forbid")


class PrincipalSummary(_Strict):
    id: str
    type: Literal["user", "group", "application"]
    display_name: str = Field(min_length=1, max_length=200)


class TenantSummary(_Strict):
    id: str
    name: str
    membership_status: Literal["invited", "active", "suspended", "removed"]


class MeResponse(_Strict):
    principal: PrincipalSummary
    current_tenant: TenantSummary
    available_tenants: list[TenantSummary]
    coarse_permissions: list[str] = Field(json_schema_extra={"uniqueItems": True})
    authorization_version: int = Field(ge=1)


class UserResponse(_Strict):
    id: str
    email: str = Field(json_schema_extra={"format": "email"})
    display_name: str
    status: Literal["active", "disabled", "deleted"]
    created_at: datetime


class MembershipResponse(_Strict):
    id: str
    user: UserResponse
    principal: PrincipalSummary
    status: Literal["invited", "active", "suspended", "removed"]
    authorization_version: int = Field(ge=1)
    created_at: datetime
    updated_at: datetime
    etag: str


class UserPage(_Strict):
    items: list[UserResponse]
    next_cursor: str | None


class MemberCandidateResponse(_Strict):
    id: str
    email: str = Field(json_schema_extra={"format": "email"})
    display_name: str


class MemberCandidatePage(_Strict):
    items: list[MemberCandidateResponse]
    next_cursor: str | None


class MembershipPage(_Strict):
    items: list[MembershipResponse]
    next_cursor: str | None


class MembershipCreate(_Strict):
    email: str = Field(min_length=3, max_length=320, json_schema_extra={"format": "email"})
    display_name: str | None = Field(default=None, max_length=200)
    role_id: UUID


class MembershipUpdate(_Strict):
    status: Literal["invited", "active", "suspended", "removed"]
    reason: str = Field(min_length=1, max_length=500)


class AddGroupMemberBody(_Strict):
    membership_id: UUID


Context = Annotated[PrincipalContext, Depends(principal_context)]


def _cursor(value: str | None, context: PrincipalContext, *, query: str) -> UUID | None:
    if value is None:
        return None
    position = CursorCodec(b"s3mp-management-cursor-key-v1").decode(
        value, context.tenant_id, context.principal_id, context.authorization_version, query=query
    )
    return UUID(position)


def _page(
    items: list[dict[str, object]], position: UUID | None, context: PrincipalContext, *, query: str
) -> dict[str, object]:
    return {
        "items": items,
        "next_cursor": CursorCodec(b"s3mp-management-cursor-key-v1").encode(
            context.tenant_id,
            context.principal_id,
            context.authorization_version,
            str(position),
            query=query,
        )
        if position
        else None,
    }


@router.get("/me", response_model=MeResponse, operation_id="get_me")
async def get_me(
    context: Context, service: Annotated[IdentityManagementService, identity_service]
) -> object:
    return await service.get_me(context)


@router.get("/users", response_model=UserPage, operation_id="list_users")
async def list_users(
    context: Annotated[PrincipalContext, management_permission("list_users")],
    service: Annotated[IdentityManagementService, identity_service],
    cursor: str | None = Query(default=None),
    status: Annotated[UserStatus, Query()] = UserStatus.ACTIVE,
    principal_type: Annotated[PrincipalType, Query()] = PrincipalType.USER,
) -> object:
    query = f"users:{status.value}:{principal_type.value}"
    items, next_position = await service.list_users(
        context,
        cursor=_cursor(cursor, context, query=query),
        status=status,
        principal_type=principal_type,
    )
    return _page(items, next_position, context, query=query)


@router.get(
    "/member-candidates",
    response_model=MemberCandidatePage,
    operation_id="list_member_candidates",
)
async def list_member_candidates(
    context: Annotated[PrincipalContext, management_permission("list_member_candidates")],
    service: Annotated[IdentityManagementService, identity_service],
    cursor: str | None = Query(default=None),
) -> object:
    # The invitation dialog loads the complete eligible candidate set and
    # filters it locally, so typing never sends personal-data search terms.
    cursor_query = "member_candidates"
    items, next_position = await service.list_member_candidates(
        context,
        cursor=_cursor(cursor, context, query=cursor_query),
    )
    return _page(items, next_position, context, query=cursor_query)


@router.get("/users/{user_id}", response_model=UserResponse, operation_id="get_user")
async def get_user(
    user_id: UUID,
    context: Annotated[PrincipalContext, management_permission("get_user")],
    service: Annotated[IdentityManagementService, identity_service],
) -> object:
    return await service.get_user(context, user_id)


@router.get("/members", response_model=MembershipPage, operation_id="list_members")
async def list_members(
    context: Annotated[PrincipalContext, management_permission("list_members")],
    service: Annotated[IdentityManagementService, identity_service],
    cursor: str | None = Query(default=None),
    status: Annotated[MembershipStatus, Query()] = MembershipStatus.ACTIVE,
) -> object:
    query = f"members:{status.value}"
    items, next_position = await service.list_members(
        context, cursor=_cursor(cursor, context, query=query), status=status
    )
    return _page(items, next_position, context, query=query)


@router.post(
    "/members", response_model=MembershipResponse, status_code=201, operation_id="create_member"
)
async def create_member(
    body: MembershipCreate,
    context: Annotated[PrincipalContext, management_permission("create_member")],
    service: Annotated[IdentityManagementService, identity_service],
    response: Response,
) -> object:
    result = await service.create_member(context, body)
    response.headers["Location"] = f"/api/v1/members/{result['id']}"
    return result


@router.get(
    "/members/{membership_id}", response_model=MembershipResponse, operation_id="get_member"
)
async def get_member(
    membership_id: UUID,
    context: Annotated[PrincipalContext, management_permission("get_member")],
    service: Annotated[IdentityManagementService, identity_service],
) -> object:
    return await service.get_member(context, membership_id)


@router.patch(
    "/members/{membership_id}", response_model=MembershipResponse, operation_id="update_member"
)
async def update_member(
    membership_id: UUID,
    body: MembershipUpdate,
    context: Annotated[PrincipalContext, management_permission("update_member")],
    service: Annotated[IdentityManagementService, identity_service],
    if_match: Annotated[str | None, Header(alias="If-Match")] = None,
) -> object:
    current = await service.get_member(context, membership_id)
    check_etag(current["etag"], require_if_match(if_match))
    return await service.update_member(context, membership_id, body)


@router.get(
    "/groups/{group_id}/members", response_model=MembershipPage, operation_id="list_group_members"
)
async def list_group_members(
    group_id: UUID,
    context: Annotated[PrincipalContext, management_permission("list_group_members")],
    service: Annotated[IdentityManagementService, identity_service],
    cursor: str | None = Query(default=None),
) -> object:
    items, next_position = await service.list_group_members(
        context, group_id, cursor=_cursor(cursor, context, query=f"group_members:{group_id}")
    )
    return _page(items, next_position, context, query=f"group_members:{group_id}")


@router.post("/groups/{group_id}/members", status_code=204, operation_id="add_group_member")
async def add_group_member(
    group_id: UUID,
    body: AddGroupMemberBody,
    context: Annotated[PrincipalContext, management_permission("add_group_member")],
    service: Annotated[IdentityManagementService, identity_service],
) -> None:
    await service.add_group_member(context, group_id, body.membership_id)


@router.delete(
    "/groups/{group_id}/members/{membership_id}",
    status_code=204,
    operation_id="remove_group_member",
)
async def remove_group_member(
    group_id: UUID,
    membership_id: UUID,
    context: Annotated[PrincipalContext, management_permission("remove_group_member")],
    service: Annotated[IdentityManagementService, identity_service],
) -> None:
    await service.remove_group_member(context, group_id, membership_id)